Show different labels for ParaTime snapshot charts
