+++ .changelog/658.trivial.md
@@ -1,0 +1,1 @@
+Update charts in ParaTime Snapshot
